Features/loans

+++ section2/loans/src/main/java/com/microservices/loans/entity/Loans.java
@@ -12,7 +12,7 @@ public class Loans extends BaseEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int loanId;
+    private Long loanId;
     private String mobileNumber;
     private String loanNumber;
     private String loanType;
